complete like function (manual and auto)

+++ project_BE/src/main/java/com/hsh/project/repository/LikeRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 @Repository
 public interface LikeRepository extends JpaRepository<Like, Integer> {
     List<Like> findByTargetTypeAndTargetId(EnumTargetType targetType, Long targetId);
-    Optional<Like> findByUserUserIdAndTargetTypeAndTargetId(Long userId, EnumTargetType targetType, Long targetId);
+    Optional<Like> findByUserUserIdAndTargetTypeAndTargetId(Integer userId, EnumTargetType targetType, Long targetId);
     @Query("SELECT COUNT(l) FROM Like l WHERE l.targetType = :targetType AND l.targetId = :targetId")
     long countByTargetTypeAndTargetId(EnumTargetType targetType, Long targetId);
 }
