+++ project_BE/src/main/java/com/hsh/project/service/spec/LikeService.java
@@ -5,7 +5,7 @@ import com.hsh.project.pojo.enums.EnumTargetType;
 import com.hsh.project.dto.response.LikeResponseDTO;
 
 public interface LikeService {
-    String toggleLike(Long userId, EnumTargetType targetType, Long targetId, EnumLikeType type);
+    String toggleLike(Integer userId, EnumTargetType targetType, Long targetId, EnumLikeType type);
     long countLikes(EnumTargetType targetType, Long targetId);
-    LikeResponseDTO createLike(Long userId, EnumTargetType targetType, Long targetId, EnumLikeType type);
+    LikeResponseDTO createLike(Integer userId, EnumTargetType targetType, Long targetId, EnumLikeType type);
 }
